tests: throughput_stats for multiple env_runners
